Proje detail
CLASSIFICATION USING THE NEAREST NEIGHBORS (KNN) ALGORITHM CLASSIFICATION ACCORDING TO THE NEAREST NEIGHBOR METHOD
There are various features extracted from real and fake banknote image samples. The image given through this information can be classified as real / fake. For each sample, 4 features (variance, skewness, kurtosis, entropy) information and whether real money (type) are readily available. We have data on a total of 1372 samples. Using this data, it is desirable to write an algorithm that finds out which of the two different types of images the image belongs to. Operations will be performed using image attributes instead of images.
a) Classification with kNN: Write down the algorithm (k nearest neighbor method) that determines which type a banknote belongs to. Write the kNN algorithm yourself, which determines which class (real (1) / fake (0)) it belongs to by taking the value of k as input for all properties of a user-entered banknote.

b) Banknote classification: Take the k value of the kNN algorithm you wrote as input and the 4 properties of a user-entered banknote as input, and list the properties, distances and classes of the nearest k banknotes on the screen as a table. Using the dataset whose connection is given on the previous page, estimate the type of the banknote with the kNN method and print it on the screen.

c) Measuring Success: Separate 100 data at the end of each banknote sample in the dataset as test data. Take the value of k from the user, classify each of the test data, using the method in item a and all 4 features, among the remaining 1172 sample data, make the listings in b. Compare the actual classes of the test data with the classes you predicted with kNN (print both real and predicted types / classes). Print the success rate as: the number of correctly classified banknotes / the total number of banknotes used for testing purposes in the dataset.

d) Listing: Type the code that displays the values in the data set in the memory.